AESP-Microservice-Project/services/chat-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
openai==1.3.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.0
elevenlabs==0.2.26
pydub==0.25.1
aiofiles==23.2.1
celery==5.3.4
pika==1.3.2
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1